refactor: rename useHomepageState.ts to use-homepage-state.ts
Standardize hook naming convention to kebab-case for consistency
with other hooks in the codebase (use-result-filters.ts).

- Rename src/hooks/useHomepageState.ts to src/hooks/use-homepage-state.ts
- Update import path in src/app/page.tsx

Fixes #42
