+++ src/hooks/use-homepage-state.ts
